Refactor plugins, utils
