+++ package.json
@@ -11,14 +11,19 @@
     "@roast-cms/react-button-beans": "*",
     "@roast-cms/react-link-filter": "*",
     "@roast-cms/react-sugar-styled": "*",
+    "@roast-cms/slate-drop-or-paste-images": "^0.8.1",
     "babel-cli": "^6.0.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
     "babel-plugin-dynamic-import-webpack": "^1.0.2",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "immutable": "^3.8.2",
+    "is-url": "^1.2.2",
+    "keycode": "^2.1.9",
     "localforage": "^1.6.0",
     "localforage-getitems": "^1.4.1",
     "lodash": "^4.17.4",
@@ -28,16 +33,12 @@
     "react-router-dom": "^4.2.2",
     "slate": "0.33.0",
     "slate-auto-replace": "^0.9.0",
-    "@roast-cms/slate-drop-or-paste-images": "^0.8.1",
     "slate-html-serializer": "^0.5.0",
     "slate-react": "0.12.0",
     "styled-components": "3.0.2",
+    "uuid": "^3.1.0",
     "webpack": "^3.11.0",
-    "webpack-dev-server": "^2.11.1",
-    "react-router-dom": "^4.2.2",
-    "keycode": "^2.1.9",
-    "is-url": "^1.2.2",
-    "uuid": "^3.1.0"
+    "webpack-dev-server": "^2.11.1"
   },
   "peerDependencies": {
     "immutable": "^3.8.2",
@@ -51,7 +52,5 @@
     "start": "webpack-dev-server",
     "prepare": "babel src --out-dir dist"
   },
-  "dependencies": {
-    "babel-plugin-transform-class-properties": "^6.24.1"
-  }
+  "dependencies": {}
 }
