+++ src/plugins/collections/README.md
@@ -1,0 +1,3 @@
+# Plugin Collections
+
+This folder groups Slate plugins by function into separate files.
