+++ src/plugins/collections/chars.js
@@ -1,0 +1,49 @@
+import AutoReplace from "slate-auto-replace"
+
+export const chars = [
+  // smart quotes
+  AutoReplace({
+    trigger: /(")/,
+    before: /[^ ”]$/,
+    transform: (transform, event, matches) => {
+      return transform.insertText("”") // smart double quote (right)
+    }
+  }),
+  AutoReplace({
+    trigger: /(")/,
+    before: /(^)|[ ]$/,
+    transform: (transform, event, matches) => {
+      return transform.insertText("“") // smart double quote (left)
+    }
+  }),
+  AutoReplace({
+    trigger: /(')/,
+    before: /[^ ”]$/,
+    transform: (transform, event, matches) => {
+      return transform.insertText("’") // smart single quote (right)
+    }
+  }),
+  AutoReplace({
+    trigger: /(')/,
+    before: /(^)|[ ]$/,
+    transform: (transform, event, matches) => {
+      return transform.insertText("‘") // smart single quote (left)
+    }
+  }),
+
+  // long dash and ellipsis
+  AutoReplace({
+    trigger: "space",
+    before: /( -)$/,
+    transform: (transform, event, matches) => {
+      return transform.insertText(" — ") // mdash
+    }
+  }),
+  AutoReplace({
+    trigger: "space",
+    before: /(\.\.\.)$/,
+    transform: (transform, event, matches) => {
+      return transform.insertText("… ") // elipsis
+    }
+  })
+]
