+++ src/utils/README.MD
@@ -1,0 +1,3 @@
+# Utility functions
+
+This folder contains functions used across the application separated by designation for better refactoring and readability.
