+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
         loader: "babel-loader",
         query: {
           presets: ["react", "es2015"],
-          plugins: ["transform-class-properties", "dynamic-import-webpack"]
+          plugins: ["transform-class-properties", "dynamic-import-webpack", "babel-plugin-transform-object-rest-spread"]
         },
         test: /\.js$/,
         exclude: /node_modules/
